+++ components/nfc/mfkey32/api/src/main/java/com/flipperdevices/nfc/mfkey32/api/MfKey32HandleDeeplink.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.nfc.mfkey32.api
+
+import android.content.Intent
+
+interface MfKey32HandleDeeplink {
+    fun handleDeepLink(intent: Intent)
+}