+++ CHANGELOG.md
@@ -10,8 +10,10 @@
 - [Feature] Check Self Update App in Options (only for github)
 - [Feature] Fap Catalog save sort
 - [Feature] Add metrics for faphub
+- [Feature] Add open mfkey32 deeplink
 - [Feature] Add dialog about failed BLE HID connection
 - [Feature] Add countly sessions 
+- [Feature] New connection REST scheme to FapHub
 - [FIX] Use by default dark theme in Wear OS
 - [FIX] Use default splashscreen in Wear OS
 - [FIX] Fix compose layout in Wear OS
@@ -28,6 +30,7 @@
 - [FIX] Openning animation
 - [FIX] Swipe refresh modifier use in root level
 - [FIX] Fix overflow in key small
+- [FIX] Try fix "The firmware update button is constantly active until the application is restarted"
 
 # 1.6.4 
 
@@ -37,7 +40,7 @@
 - [Feature] Add fap manifest caching
 - [Feature] Not send empty report in faphub
 - [Feature] Add hide app button
-- [Feature] Check Self Update App in Optiond (only for github)
+- [Feature] Check Self Update App in Option (only for github)
 - [CI] Add apk artifacts to Pull Requests
 - [CI] Fix failed build in merge_group
 - [CI] Bump deps except Sentry-gradle